Optimize allocation strategy
Most of the program runtime was spent in malloc(), so we add a simplistic
memory pool. We allocate a big chunk of objects at once and store them in a
linked list. Then deallocation is two pointer assignments and allocation is
that plus a branch (unless we actually run out of memory, when we allocate a
bunch more objects in one fell swoop).

Program performance on a release build is improved ~4x by my measurements.